requirements.txt
uvicorn
fastapi
torchaudio